allow Length[freehand function] and First[freehand function, n] (Fix for Locus, add test)
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@52756 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/kernel/algos/AlgoFunctionFreehand.java
@@ -116,7 +116,7 @@ public class AlgoFunctionFreehand extends AlgoElement {
 		return inputList;
 	}
 
-	public double getPointLength() {
+	public int getPointLength() {
 		// first two elements are the min/max so subtract 2
 		return inputList.size() - 2;
 	}
